backend/requirements.txt
# Backend Dependencies
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1
pydantic-settings==2.1.0
httpx==0.26.0
python-multipart==0.0.6
